Implement new LED driver
Verification for the new LED driver is completed so time to replace the
previous gpio implementation with the new one. This entails
replacing the led_controller file with the led_controller_new file,
updating the dbus interface name so that it is the same as the previous
one, and updating the code with the new led names.
